requirements.txt
asgiref==3.4.1
certifi==2021.5.30
cycler==0.11.0
Django==3.2.9
django-cleanup==5.2.0
django-extensions==3.1.5
django-nested-admin==3.3.3
joblib==1.1.0
kiwisolver==1.3.1
matplotlib==3.3.4
numpy==1.19.5
pandas==1.1.5
Pillow==8.4.0
pyparsing==3.0.6
python-dateutil==2.8.2
python-monkey-business==1.0.0
pytz==2021.3
scikit-learn==0.22.1
scipy==1.5.4
six==1.16.0
sqlparse==0.4.2
threadpoolctl==3.0.0
typing_extensions==4.0.1
